pyproject.toml
[project]
name = "inmueblesapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.5",
    "ollama>=0.5.3",
    "optuna>=4.5.0",
    "pandas>=2.3.1",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "tensorflow>=2.20.0",
    "tqdm>=4.67.1",
    "xgboost>=3.0.4",
]
